requirements.txt
Django==5.1.7
djangorestframework==3.15.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.6
django-celery-results==2.5.1
PyJWT==2.8.0
bcrypt==4.1.2
pandas==2.2.1
dj-database-url==2.3.0
